Make cpe_dict_model.origin_file accessors private
