+++ src/CPE/cpedict_priv.c
@@ -1288,6 +1288,19 @@ static void cpe_notes_export(const struct cpe_notes *notes, xmlTextWriterPtr wri
 	xmlTextWriterEndElement(writer);
 }
 
+bool cpe_dict_model_set_origin_file(struct cpe_dict_model* dict, const char* origin_file)
+{
+	free(dict->origin_file);
+	dict->origin_file = oscap_strdup(origin_file);
+
+	return true;
+}
+
+const char* cpe_dict_model_get_origin_file(const struct cpe_dict_model* dict)
+{
+	return dict->origin_file;
+}
+
 /* End of private export functions
  * */
 /***************************************************************************/
